packages/openai_agents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anosys-sdk-openai-agents"
version = "1.0.0"
description = "AnoSys SDK for OpenAI Agents - Automatic instrumentation and logging for OpenAI Agents"
authors = [{name = "AnoSys", email = "support@anosys.ai"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
keywords = ["anosys", "openai", "agents", "observability", "monitoring", "logging", "llm"]
dependencies = [
    "anosys-sdk-core>=1.0.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "traceai-openai-agents>=0.1.0",
    "openai-agents>=0.0.1",
]

[project.urls]
Homepage = "https://anosys.ai"
Documentation = "https://docs.anosys.ai"
Repository = "https://github.com/anosys-ai/anosys-sdk"

[tool.setuptools.packages.find]
where = ["src"]
